+++ frontend/tailwind.config.js
@@ -6,16 +6,16 @@ module.exports = {
     './components/**/*.{js,ts,jsx,tsx}',
   ],
   theme: {
-    container: {
-      center: true,
-      padding: {
-        DEFAULT: '1rem',
-        sm: '2rem',
-        lg: '4rem',
-        xl: '5rem',
-        '2xl': '6rem',
-      },
-    },
+    // container: {
+    //   center: true,
+    //   padding: {
+    //     DEFAULT: '1rem',
+    //     sm: '2rem',
+    //     lg: '4rem',
+    //     xl: '5rem',
+    //     '2xl': '6rem',
+    //   },
+    // },
     extend: {
       fontFamily: {
         sans: ['Dosis', ...defaultTheme.fontFamily.sans],
